Only delete schedule it finds

+++ app/Jobs/ScheduleVSTPCreate.php
@@ -42,7 +42,9 @@ class ScheduleVSTPCreate implements ShouldQueue
                                           ->where('start_date', 'like', $this->schedule->schedule_start_date)
                                           ->where('stp_indicator', 'like', $this->schedule->CIF_stp_indicator)
                                           ->first();
-            $expiredSchedule->delete();                              
+            if(!is_null($expiredSchedule)){
+                $expiredSchedule->delete();
+            }
 
         } else if(strtolower($this->schedule->transaction_type) == "create") {
